+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 - CHANGED: Updated whois.nic.af parser to the new response format.
 
+- CHANGED: Updated whois.nic.cx parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).